reader_sage/pyproject.toml
[tool.poetry]
name = "reader-sage"
version = "0.1.0"
description = ""
authors = ["Константин <k1st.volkov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
mysql-connector-python = "^8.3.0"
python-telegram-bot = "^20.8"
fastapi = {extras = ["all"], version = "^0.110.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
